Update phaser_gene_ae.py
Added "--no_gw_phase" argument which allows phaser_gene_ae to be run on samples that were not whole genome phased before phASER was run. When running in this mode only the SNP or haplotype block with the highest coverage per gene is used, and phase between genes is not preserved.