Version 0.5 Update - phaser.py
Update with major improvements to haplotype assembly algorithm, resulting in much faster runtimes. In our testing runtime was between 2x and 50x faster than the previous version. Note the new version requires the use of an external script (read_variant_map.py), so this must be in the same directory as the main script (phaser.py). Please also ensure that if python is called with a command other than "python2.7" that you set this option using the "--python_string" argument.