Fix: Parse AS tag only when string starts with "AS:"
I have a BAM file (Genome in a Bottle data) where there are some extra tags added to each record that seem to encode quality values of some type. Some of these quality strings actually contain the string `AS:` just by chance. Without this fix, phaser mistakenly recognizes that quality tag as the one for the alignment score and crashes when it tries to intepret the part after it as an integer.