v1.1.4
phASER Gene AE now requires that at least one variant of a haplotype block is found within a gene annotation in order for that block's read counts to be assigned to the gene. Previously, if a haplotype block simply spanned a gene annotation its counts would get included for that gene, even if none of the variants in the block were themselves located within the gene. In cases where a gene is located inside of another gene, e.g. in the intron of another gene, this would result in the gene getting assigned counts of the gene it is located in.

We are planning a future update that will only use read counts from variants that are found within the gene annotation region, which should result in better accuracy in cases where the genomic coordinates of distinct transcripts overlap one another.